backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
snowflake-connector-python>=3.6.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
PyJWT>=2.8.0
cryptography>=41.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
httpx>=0.26.0
